+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/notifications/handler/ChatNotificationHandler.kt
@@ -114,7 +114,7 @@ public open class ChatNotificationHandler @JvmOverloads constructor(
         ).build()
     }
 
-    public open fun showNotification(channel: Channel, message: Message) {
+    internal fun showNotification(channel: Channel, message: Message) {
         val notificationId: Int = System.nanoTime().toInt()
         val notificationSummaryId = getNotificationGroupSummaryId(channel.type, channel.id)
         addNotificationId(notificationId, notificationSummaryId)
@@ -185,6 +185,16 @@ public open class ChatNotificationHandler @JvmOverloads constructor(
         return context.packageManager!!.getLaunchIntentForPackage(context.packageName)!!
     }
 
+    /**
+     * Dismiss notifications from a given [channelType] and [channelId].
+     *
+     * @param channelType String that represent the channel type of the channel you want to dismiss notifications.
+     * @param channelId String that represent the channel id of the channel you want to dismiss notifications.
+     */
+    internal fun dismissChannelNotifications(channelType: String, channelId: String) {
+        dismissSummaryNotification(getNotificationGroupSummaryId(channelType, channelId))
+    }
+
     public open fun getErrorCaseIntent(): Intent {
         return context.packageManager!!.getLaunchIntentForPackage(context.packageName)!!
     }
@@ -232,6 +242,15 @@ public open class ChatNotificationHandler @JvmOverloads constructor(
     private fun Channel.getMemberNamesWithoutCurrentUser(): String? = getUsersExcludingCurrent()
         .joinToString { it.name }
         .takeIf { it.isNotEmpty() }
+
+    private fun dismissSummaryNotification(notificationSummaryId: Int) {
+        getAssociatedNotificationIds(notificationSummaryId).forEach {
+            notificationManager.cancel(it)
+            removeNotificationId(it)
+        }
+        notificationManager.cancel(notificationSummaryId)
+        sharedPreferences.edit { remove(getNotificationSummaryIdKey(notificationSummaryId)) }
+    }
 
     internal fun onDismissNotification(notificationId: Int) {
         val notificationSummaryId = getAssociatedNotificationSummaryId(notificationId)