+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added a method to dismiss all notifications from a channel. It is handled internally from the SDK but you are able to dismiss channel notification at whatever time calling `ChatClient::dismissChannelNotifications`
 
 ### ⚠️ Changed
 
@@ -68,6 +69,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Notifications are dismissed after the user go into the channel conversation when you are using `MessageListView`
 
 ### ⚠️ Changed
 
@@ -82,6 +84,7 @@
 - Updated the Accompanist library version (0.18.0)
 
 ### ✅ Added
+- Notifications are dismissed after the user go into the channel conversation when you are using `MessageList`
 
 ### ⚠️ Changed